refactor: rename data -> constants
