+++ src/components/Carousel/CustomCarousel.tsx
@@ -10,7 +10,7 @@ import {
   CarouselPrevious,
 } from '@/components/ui/carousel'
 import ShoppingCard from '../Cards/ShoppingCard'
-import { shoppingData } from '@/data/sectionData'
+import { shoppingData } from '@/constants/sectionData'
 
 type CarouselProps = {
   category: string
